hide the faulty half-sized background shade
seems like overkill forgot the second half (right-side) of this background.
it's hard to see, but when you notice it once, you will see it all the time.

adding the missing part would probably be the better looking option, but then we would also need to rewrite the nested local functions in animate_eye(), which would be quite a mess.
so let's just get rid of the incomplete background instead.